build(lint): enforce clang-tidy-20 and clang-format-20, and apply the first pass
The tree adopts clang-tidy-20 with a check set tuned for the legacy
codebase (bugprone, performance, and a narrow modernization set) and
clang-format-20 enforcement in CI, and the first pass is applied
across the tree: formatting throughout, plus mechanical,
behavior-equivalent modernizations the linters require (NULL to
nullptr conversions and NOLINT suppressions where a finding is
accepted). The pass is recorded in .git-blame-ignore-revs so blame
stays useful across it.

+++ net/instaweb/rewriter/webp_optimizer.cc
@@ -195,7 +195,9 @@ bool WebpOptimizer::DoReadJpegPixels(J_COLOR_SPACE color_space,
   // Figure out critical dimensions of image, and allocate space for image data.
   width_ = jpeg_decompress->output_width;
   height_ = jpeg_decompress->output_height;
-  row_stride_ = width_ * jpeg_decompress->output_components * sizeof(*pixels_);
+  row_stride_ =
+      static_cast<unsigned long>(width_ * jpeg_decompress->output_components) *
+      sizeof(*pixels_);
 
   pixels_ = new uint8[row_stride_ * height_];
   // jpeglib expects to get an array of pointers to rows, so allocate one and
